% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_wrapper.R
\name{global_wrapper}
\alias{global_wrapper}
\title{Title}
\usage{
global_wrapper(
  par,
  fn,
  lower,
  upper,
  method,
  print = TRUE,
  control = list(),
  ...
)
}
\arguments{
\item{par}{Vector. Initial values for the components to be optimized.}

\item{fn}{A function to be minimized, with first argument the vector of parameters over which minimization is to take place}

\item{lower}{Vector with length of par. Lower bounds for components}

\item{upper}{Vector with length of par. Upper bounds for components}

\item{method}{The method to be used}

\item{control}{The argument is a list that can be used to control the behavior of the algorithm}

\item{...}{allows the user to pass additional arguments to the function fn}
}
\value{
ans The object returned by the optimizer
}
\description{
Title
}
