% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochOptim_method_list.R
\name{StochOptim_method_list}
\alias{StochOptim_method_list}
\title{A List of all the optimizer in StochOptim}
\description{
A List of all the optimizer in StochOptim
}
\section{Methods}{

the list of methods are :-
\itemize{
\item DEoptim
\item pso
\item GenSA
\item DEoptimR
\item adagio::simpleDE,
\item adagio::pureCMAES
\item NMOF::DEopt
\item NMOF::GAopt
\item NMOF::PSopt,
\item nloptr::isres
\item nloptr::stogo
\item cmaes::cam_es
\item soma
\item ceimOpt
\item RcppDE::DEoptim
}
}

