% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sopm.R
\name{sopm}
\alias{sopm}
\title{Calls multiple optimizers}
\usage{
sopm(par, fn, lower, upper, method, control = list(), ...)
}
\arguments{
\item{par}{Vector. Initial values for the components to be optimized.}

\item{fn}{A function to be minimized, with first argument the vector of parameters over which minimization is to take place}

\item{lower}{Vector with length of par. Lower bounds for components}

\item{upper}{Vector with length of par. Upper bounds for components}

\item{method}{The method to be used}

\item{control}{The argument is a list that can be used to control the behavior of the algorithm}

\item{...}{allows the user to pass additional arguments to the function fn}
}
\value{
A dataframe containing values returned by optimizer.
}
\description{
Calls multiple optimizers
}
\examples{
###################################################

# simple demonstration without any controls
fn <- function(x, a = 10) return(sum(x^2) + a)
dim <- 10
par <- rnorm(dim)
lb <- rep(-20,dim)
ub <- rep(20, dim)
res <- sopm(par = par, fn = fn, lower = lb, upper = ub,
           method = c("DEoptim", "DEoptimR", "GenSA",
           "pso", "adagio_simpleDE"))
res
##################################################

# passing control parameters which are common between methods.

# here maxiter is common between solvers but since GenSA specifically
# defines maxiter for itself it takes precedence. Hence maxiter for DEoptim
will be 100 but for GenSA it will be 50

control <- list(maxiter = 100,
DEoptim = list(tol = 1e-10), GenSA = list(maxiter = 50))
sopm(par = par, fn = fn,lower = lb, upper = ub,
method = c("DEoptim","GenSA") ,control = control)
###################################################

# Passing control parameter individually to each method

fn <- rastrigin <- function(x) {
y <- ( 10 * length(x) + sum(x^2 - 10 * cos(2 * pi * x)))
return(y)
}

D <- 10
lower <- lb <- rep(-5.2, D)
upper <- ub <- rep(5.2, D)
par <- rep(3.3, D)

DEoptim_list = list(
popsize = 10*length(lb),
maxiter = 200*length(lb),
tol  = 1e-10,
trace   = 0)

adagio_simpleDE_list = list(
 popsize  = 70,
 nmax     = 230,
 r        = 0.41,
 confined = TRUE,
 trace    = FALSE)

GenSA_list  = list(maxiter = 100)
pso_list    = list(abstol= 1e-8, hybrid = "improved")
DEoptimR_list = list(tol = 1e-7, trace = FALSE)

control <- list(
 DEoptim = DEoptim_list,
 GenSA = GenSA_list,
 pso = pso_list,
 DEoptimR  = DEoptimR_list,
 adagio_simpleDE = adagio_simpleDE_list)

sopm(par = par,
    fn = fn,
    lower = lb,
    upper = ub,
    method = c("DEoptim", "GenSA","pso","DEoptimR","adagio_simpleDE"),
    control= control)
}
